+++ CHANGE_LOG.md
@@ -1,6 +1,7 @@
 # v5.8.7
 - Support .NET8
 - Add vkey and lkey to license file
+- Add X509 license services
 
 # v5.8.6
 - Distuiguish between oid and sub claims for license